Break safir.gcs up into internal modules
Move the code for safir.gcs out of __init__.py into internal
modules and turn __init__.py into only imports and re-exports.
